Creates graph and parsing code instantiation code
This class will contain a graph representation and code for flattening the chain to get all dependencies.

It's not clear yet how the class split will be: there may be a separate class for parsing from the list of strings and for the graph representation, and possibly even for the flattening of the graph itself.