Creates linked list merge algorithm
This implementation must be immediately updated to use a non-.NET implementation of a linked list. Firstly, the default linked list is unnecessarily (for the purpose of the algorithm) a doubly-linked list.
Moreover, the linked list nodes have non-writable Next properties. Microsoft probably did this to ensure that lower level linked list properties and behaviours are encapsulated, thereby ensuring that the list is always in a valid state.

There is also a limitation in the merge algorithm in that information is being lost when the inner and outer node have the same values and are the outer node is simply skipped (outerNode = outerNode.Next). If these skipped nodes were tracked, it'd be easier to avoid having to do more merges during the DFS search. The collection of skipped nodes would then have to be accessibly to the parent algorithm - better performance could be obtained by not decoupling the code two separate methods in this way.