Adds tests for singly linked list and merging
These are needed for the internal testing of linked lists. The tests don't need to cover all cases for linked list merging - just the case for when they are sorted. This reduction in scope should mean that the linked list implementation and required algorithms can be written for the required scenario (graph representation and _sorted_ linked list merging)