Adds a graph bug fixes and adds GetDependencyListOfAllVertexDependencies
This includes a new flattened hashset for determining when a vertex has had all dependents flattened into it's linked list chain. Also adds a few bug fixes related to merging and skipping visiting certain dependents.

Moreover, the GetDependencyListOfAllVertexDependencies method gets a list of the flattened chains, which will be the output from the graph flattening process.